strategy_live/requirements.txt
Connect==0.2
mysql_connector_repackaged==0.3.1
pandas==1.5.3
python-socketio==5.11.2
python_dateutil==2.8.2
Requests==2.32.3
